requirements.txt
Django==4.2
tensorflow==2.12.0
numpy
pillow
python-dotenv
dnspython
pymongo
